Overhaul PythonAPI, switch to scikit-build-core.
