+++ LibCarla/source/carla/Memory.h
@@ -11,9 +11,6 @@
 
 namespace carla {
 
-  /// Use this SharedPtr (std::shared_ptr) to keep compatibility with
-  /// boost::python, but it would be nice if in the future we can make a Python
-  /// adaptor for std::shared_ptr.
   template <typename T>
   using SharedPtr = std::shared_ptr<T>;
