+++ PythonAPI/carla/source/libcarla/Client.cpp
@@ -4,12 +4,7 @@
 // This work is licensed under the terms of the MIT license.
 // For a copy, see <https://opensource.org/licenses/MIT>.
 
-#include "carla/PythonUtil.h"
-#include "carla/client/Client.h"
-#include "carla/client/World.h"
-#include "carla/Logging.h"
-#include "carla/rpc/ActorId.h"
-#include "carla/trafficmanager/TrafficManager.h"
+#include "PythonAPI.h"
 
 #include <thread>
