+++ PythonAPI/carla/source/libcarla/Control.cpp
@@ -5,129 +5,9 @@
 // This work is licensed under the terms of the MIT license.
 // For a copy, see <https://opensource.org/licenses/MIT>.
 
-#include <carla/rpc/VehicleAckermannControl.h>
-#include <carla/rpc/VehicleControl.h>
-#include <carla/rpc/VehiclePhysicsControl.h>
-#include <carla/rpc/WheelPhysicsControl.h>
-#include <carla/rpc/WalkerControl.h>
-#include <carla/rpc/WalkerBoneControlIn.h>
-#include <carla/rpc/WalkerBoneControlOut.h>
+#include "PythonAPI.h"
 
-#include <ostream>
-
-namespace carla {
-namespace rpc {
-
-  static auto boolalpha(bool b) {
-    return b ? "True" : "False";
-  }
-
-  std::ostream &operator<<(std::ostream &out, const VehicleControl &control) {
-    out << "VehicleControl(throttle=" << std::to_string(control.throttle)
-        << ", steer=" << std::to_string(control.steer)
-        << ", brake=" << std::to_string(control.brake)
-        << ", hand_brake=" << boolalpha(control.hand_brake)
-        << ", reverse=" << boolalpha(control.reverse)
-        << ", manual_gear_shift=" << boolalpha(control.manual_gear_shift)
-        << ", gear=" << std::to_string(control.gear) << ')';
-    return out;
-  }
-
-  std::ostream &operator<<(std::ostream &out, const VehicleAckermannControl &control) {
-    out << "VehicleAckermannControl(steer=" << std::to_string(control.steer)
-        << ", steer_speed=" << std::to_string(control.steer_speed)
-        << ", speed=" << std::to_string(control.speed)
-        << ", acceleration=" << std::to_string(control.acceleration)
-        << ", jerk=" << std::to_string(control.jerk) << ')';
-    return out;
-  }
-
-  std::ostream &operator<<(std::ostream &out, const WalkerControl &control) {
-    out << "WalkerControl(direction=" << control.direction
-        << ", speed=" << std::to_string(control.speed)
-        << ", jump=" << boolalpha(control.jump) << ')';
-    return out;
-  }
-
-  std::ostream &operator<<(std::ostream &out, const WalkerBoneControlIn &control) {
-    out << "WalkerBoneControlIn(bone_transforms(";
-    for (auto bone_transform : control.bone_transforms) {
-      out << "(name="  << bone_transform.first
-          << ", transform=" << bone_transform.second << ')';
-    }
-    out << "))";
-    return out;
-  }
-
-  std::ostream &operator<<(std::ostream &out, const BoneTransformDataOut &data) {
-    out << "BoneTransformDataOut(name="  << data.bone_name << ", world=" << data.world << ", component=" << data.component << ", relative=" << data.relative << ')';
-    return out;
-  }
-
-  std::ostream &operator<<(std::ostream &out, const WalkerBoneControlOut &control) {
-    out << "WalkerBoneControlOut(bone_transforms(";
-    for (auto bone_transform : control.bone_transforms) {
-      out << "(name="  << bone_transform.bone_name
-          << ", world=" << bone_transform.world << ", component=" << bone_transform.component << ", relative=" << bone_transform.relative << ')';
-    }
-    out << "))";
-    return out;
-  }
-
-  std::ostream &operator<<(std::ostream &out, const GearPhysicsControl &control) {
-    out << "GearPhysicsControl(ratio=" << std::to_string(control.ratio)
-        << ", down_ratio=" << std::to_string(control.down_ratio)
-        << ", up_ratio=" << std::to_string(control.up_ratio) << ')';
-    return out;
-  }
-
-  std::ostream &operator<<(std::ostream &out, const WheelPhysicsControl &control) {
-    out << "WheelPhysicsControl(tire_friction=" << std::to_string(control.tire_friction)
-        << ", damping_rate=" << std::to_string(control.damping_rate)
-        << ", max_steer_angle=" << std::to_string(control.max_steer_angle)
-        << ", radius=" << std::to_string(control.radius)
-        << ", max_brake_torque=" << std::to_string(control.max_brake_torque)
-        << ", max_handbrake_torque=" << std::to_string(control.max_handbrake_torque)
-        << ", lat_stiff_max_load=" << std::to_string(control.lat_stiff_max_load)
-        << ", lat_stiff_value=" << std::to_string(control.lat_stiff_value)
-        << ", long_stiff_value=" << std::to_string(control.long_stiff_value)
-        << ", position=" << control.position << ')';
-    return out;
-  }
-
-  std::ostream &operator<<(std::ostream &out, const VehiclePhysicsControl &control) {
-    out << "VehiclePhysicsControl(torque_curve=" << control.torque_curve
-    << ", max_rpm=" << std::to_string(control.max_rpm)
-    << ", moi=" << std::to_string(control.moi)
-    << ", damping_rate_full_throttle=" << std::to_string(control.damping_rate_full_throttle)
-    << ", damping_rate_zero_throttle_clutch_engaged=" << std::to_string(control.damping_rate_zero_throttle_clutch_engaged)
-    << ", damping_rate_zero_throttle_clutch_disengaged=" << std::to_string(control.damping_rate_zero_throttle_clutch_disengaged)
-    << ", use_gear_autobox=" << boolalpha(control.use_gear_autobox)
-    << ", gear_switch_time=" << std::to_string(control.gear_switch_time)
-    << ", clutch_strength=" << std::to_string(control.clutch_strength)
-    << ", final_ratio=" << std::to_string(control.final_ratio)
-    << ", forward_gears=" << control.forward_gears
-    << ", mass=" << std::to_string(control.mass)
-    << ", drag_coefficient=" << std::to_string(control.drag_coefficient)
-    << ", center_of_mass=" << control.center_of_mass
-    << ", steering_curve=" << control.steering_curve
-    << ", wheels=" << control.wheels
-    << ", use_sweep_wheel_collision=" << control.use_sweep_wheel_collision << ')';
-    return out;
-  }
-
-  std::ostream &operator<<(std::ostream &out, const AckermannControllerSettings &settings) {
-    out << "AckermannControllerSettings(speed_kp=" << std::to_string(settings.speed_kp)
-        << ", speed_ki=" << std::to_string(settings.speed_ki)
-        << ", speed_kd=" << std::to_string(settings.speed_kd)
-        << ", accel_kp=" << std::to_string(settings.accel_kp)
-        << ", accel_ki=" << std::to_string(settings.accel_ki)
-        << ", accel_kd=" << std::to_string(settings.accel_kd)  << ')';
-    return out;
-  }
-
-} // namespace rpc
-} // namespace carla
+#include <boost/python/suite/indexing/vector_indexing_suite.hpp>
 
 static auto GetVectorOfVector2DFromList(const boost::python::list &list) {
   std::vector<carla::geom::Vector2D> v;
