+++ PythonAPI/carla/source/libcarla/Sensor.cpp
@@ -4,11 +4,7 @@
 // This work is licensed under the terms of the MIT license.
 // For a copy, see <https://opensource.org/licenses/MIT>.
 
-#include <carla/PythonUtil.h>
-#include <carla/client/ClientSideSensor.h>
-#include <carla/client/LaneInvasionSensor.h>
-#include <carla/client/Sensor.h>
-#include <carla/client/ServerSideSensor.h>
+#include "PythonAPI.h"
 
 static void SubscribeToStream(carla::client::Sensor &self, boost::python::object callback) {
   self.Listen(MakeCallback(std::move(callback)));
