+++ PythonAPI/carla/source/libcarla/TrafficManager.cpp
@@ -4,14 +4,8 @@
 // This work is licensed under the terms of the MIT license.
 // For a copy, see <https://opensource.org/licenses/MIT>.
 
-#include <chrono>
-#include <memory>
-#include <stdio.h>
-#include "carla/PythonUtil.h"
+#include "PythonAPI.h"
 #include "boost/python/suite/indexing/vector_indexing_suite.hpp"
-
-#include "carla/trafficmanager/TrafficManager.h"
-#include "carla/trafficmanager/SimpleWaypoint.h"
 
 using ActorPtr = carla::SharedPtr<carla::client::Actor>;
 using ActorId = carla::ActorId;
