+++ PythonAPI/carla/source/libcarla/Weather.cpp
@@ -4,9 +4,7 @@
 // This work is licensed under the terms of the MIT license.
 // For a copy, see <https://opensource.org/licenses/MIT>.
 
-#include <carla/rpc/WeatherParameters.h>
-
-#include <ostream>
+#include "PythonAPI.h"
 
 namespace carla {
 namespace rpc {
